Move files from `app/lib/pages` to `app/lib/blackboard`
