+++ app/lib/blackboard/blackboard_page.dart
@@ -10,12 +10,12 @@ import 'package:analytics/analytics.dart';
 import 'package:bloc_provider/bloc_provider.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_staggered_animations/flutter_staggered_animations.dart';
+import 'package:sharezone/blackboard/blackboard_dialog.dart';
 import 'package:sharezone/blocs/application_bloc.dart';
 import 'package:sharezone/blocs/blackboard/blackboard_page_bloc.dart';
 import 'package:sharezone/navigation/logic/navigation_bloc.dart';
 import 'package:sharezone/navigation/models/navigation_item.dart';
 import 'package:sharezone/navigation/scaffold/sharezone_main_scaffold.dart';
-import 'package:sharezone/pages/blackboard/blackboard_dialog.dart';
 import 'package:sharezone/widgets/blackboard/blackboard_card.dart';
 import 'package:sharezone/widgets/blackboard/blackboard_view.dart';
 import 'package:sharezone_widgets/placeholder.dart';
@@ -23,8 +23,8 @@ import 'package:sharezone_widgets/snackbars.dart';
 import 'package:sharezone_widgets/theme.dart';
 import 'package:sharezone_widgets/widgets.dart';
 
-import 'blackboard/blackboard_dialog.dart';
-import 'blackboard/details/blackboard_details.dart';
+import 'blackboard_dialog.dart';
+import 'details/blackboard_details.dart';
 
 /// Open the blackboard page and returns true, if the user added a blackboard item
 Future<bool> openBlackboardDialogAndShowConfirmationIfSuccessful(
