+++ app/lib/blackboard/details/blackboard_details_bloc.dart
