+++ app/lib/blackboard/details/blackboard_item_read_by_users_list/blackboard_item_read_by_users_list_page.dart
@@ -9,8 +9,8 @@
 import 'package:bloc_provider/bloc_provider.dart';
 import 'package:common_domain_models/common_domain_models.dart';
 import 'package:flutter/material.dart';
+import 'package:sharezone/blackboard/analytics/blackboard_analytics.dart';
 import 'package:sharezone/blocs/application_bloc.dart';
-import 'package:sharezone/pages/blackboard/analytics/blackboard_analytics.dart';
 import 'package:sharezone_widgets/additional.dart';
 import 'package:sharezone_widgets/wrapper.dart';
 import 'package:user/user.dart';
