+++ app/lib/blackboard/details/blackboard_item_read_by_users_list/user_view.dart
